+++ src/lib.rs
@@ -15,8 +15,10 @@ use crate::{
 };
 use anyhow::Context as _;
 use cargo_metadata as cm;
+use itertools::{iproduct, Itertools as _};
+use krates::PkgSpec;
 use maplit::hashset;
-use std::{cmp, path::PathBuf, str::FromStr};
+use std::{cmp, collections::BTreeMap, path::PathBuf, str::FromStr};
 use structopt::{clap::AppSettings, StructOpt};
 
 #[derive(StructOpt, Debug)]
@@ -48,6 +50,18 @@ pub enum Opt {
         /// Path to Cargo.toml
         #[structopt(long, value_name("PATH"))]
         manifest_path: Option<PathBuf>,
+
+        /// Exclude library crates from bundling
+        #[structopt(long, value_name("SPEC"))]
+        exclude: Vec<PkgSpec>,
+
+        /// Alias for `--exclude https://github.com/rust-lang/crates.io-index#alga:0.9.3 ..`
+        #[structopt(long)]
+        exclude_atcoder_crates: bool,
+
+        /// Alias for `--exclude https://github.com/rust-lang/crates.io-index#chrono:0.4.9 ..`
+        #[structopt(long)]
+        exclude_codingame_crates: bool,
 
         /// `nightly` toolchain for `cargo-udeps`
         #[structopt(long, value_name("TOOLCHAIN"), default_value("nightly"))]
@@ -145,10 +159,64 @@ pub struct Context<'a> {
 }
 
 pub fn run(opt: Opt, ctx: Context<'_>) -> anyhow::Result<()> {
+    static ATCODER_CRATES: &[&str] = &[
+        "https://github.com/rust-lang/crates.io-index#alga:0.9.3",
+        "https://github.com/rust-lang/crates.io-index#ascii:1.0.0",
+        "https://github.com/rust-lang/crates.io-index#bitset-fixed:0.1.0",
+        "https://github.com/rust-lang/crates.io-index#either:1.5.3",
+        "https://github.com/rust-lang/crates.io-index#fixedbitset:0.2.0",
+        "https://github.com/rust-lang/crates.io-index#getrandom:0.1.14",
+        "https://github.com/rust-lang/crates.io-index#im-rc:14.3.0",
+        "https://github.com/rust-lang/crates.io-index#indexmap:1.3.2",
+        "https://github.com/rust-lang/crates.io-index#itertools:0.9.0",
+        "https://github.com/rust-lang/crates.io-index#itertools-num:0.1.3",
+        "https://github.com/rust-lang/crates.io-index#lazy_static:1.4.0",
+        "https://github.com/rust-lang/crates.io-index#libm:0.2.1",
+        "https://github.com/rust-lang/crates.io-index#maplit:1.0.2",
+        "https://github.com/rust-lang/crates.io-index#nalgebra:0.20.0",
+        "https://github.com/rust-lang/crates.io-index#ndarray:0.13.0",
+        "https://github.com/rust-lang/crates.io-index#num:0.2.1",
+        "https://github.com/rust-lang/crates.io-index#num-bigint:0.2.6",
+        "https://github.com/rust-lang/crates.io-index#num-complex:0.2.4",
+        "https://github.com/rust-lang/crates.io-index#num-derive:0.3.0",
+        "https://github.com/rust-lang/crates.io-index#num-integer:0.1.42",
+        "https://github.com/rust-lang/crates.io-index#num-iter:0.1.40",
+        "https://github.com/rust-lang/crates.io-index#num-rational:0.2.4",
+        "https://github.com/rust-lang/crates.io-index#num-traits:0.2.11",
+        "https://github.com/rust-lang/crates.io-index#ordered-float:1.0.2",
+        "https://github.com/rust-lang/crates.io-index#permutohedron:0.2.4",
+        "https://github.com/rust-lang/crates.io-index#petgraph:0.5.0",
+        "https://github.com/rust-lang/crates.io-index#proconio:0.3.6",
+        "https://github.com/rust-lang/crates.io-index#rand:0.7.3",
+        "https://github.com/rust-lang/crates.io-index#rand_chacha:0.2.2",
+        "https://github.com/rust-lang/crates.io-index#rand_core:0.5.1",
+        "https://github.com/rust-lang/crates.io-index#rand_distr:0.2.2",
+        "https://github.com/rust-lang/crates.io-index#rand_hc:0.2.0",
+        "https://github.com/rust-lang/crates.io-index#rand_pcg:0.2.1",
+        "https://github.com/rust-lang/crates.io-index#regex:1.3.6",
+        "https://github.com/rust-lang/crates.io-index#rustc-hash:1.1.0",
+        "https://github.com/rust-lang/crates.io-index#smallvec:1.2.0",
+        "https://github.com/rust-lang/crates.io-index#superslice:1.0.0",
+        "https://github.com/rust-lang/crates.io-index#text_io:0.1.8",
+        "https://github.com/rust-lang/crates.io-index#whiteread:0.5.0",
+    ];
+
+    static CODINGAME_CRATES: &[&str] = &[
+        "https://github.com/rust-lang/crates.io-index#chrono:0.4.9",
+        "https://github.com/rust-lang/crates.io-index#itertools:0.8.0",
+        "https://github.com/rust-lang/crates.io-index#libc:0.2.62",
+        "https://github.com/rust-lang/crates.io-index#rand:0.7.2",
+        "https://github.com/rust-lang/crates.io-index#regex:1.3.0",
+        "https://github.com/rust-lang/crates.io-index#time:0.1.42",
+    ];
+
     let Opt::Equip {
         src,
         bin,
         manifest_path,
+        exclude,
+        exclude_atcoder_crates,
+        exclude_codingame_crates,
         toolchain,
         resolve_cfgs,
         remove,
@@ -162,6 +230,17 @@ pub fn run(opt: Opt, ctx: Context<'_>) -> anyhow::Result<()> {
     let minify = match (minify, oneline) {
         (Minify::None, oneline) => oneline,
         (minify, _) => minify,
+    };
+
+    let exclude = {
+        let mut exclude = exclude;
+        if exclude_atcoder_crates {
+            exclude.extend(ATCODER_CRATES.iter().map(|s| s.parse().unwrap()));
+        }
+        if exclude_codingame_crates {
+            exclude.extend(CODINGAME_CRATES.iter().map(|s| s.parse().unwrap()));
+        }
+        exclude
     };
 
     let Context { cwd, shell } = ctx;
@@ -182,20 +261,69 @@ pub fn run(opt: Opt, ctx: Context<'_>) -> anyhow::Result<()> {
         metadata.exactly_one_bin_target()
     }?;
 
+    let deps_to_bundle = {
+        let unused_deps = match cargo_udeps::cargo_udeps(&bin_package, &bin.name, &toolchain, shell)
+        {
+            Ok(unused_deps) => unused_deps,
+            Err(warning) => {
+                shell.warn(warning)?;
+                hashset!()
+            }
+        };
+        metadata.deps_to_bundle(&bin_package.id, &unused_deps, &exclude)?
+    };
+
+    let error_message = |head: &str| {
+        let mut msg = head.to_owned();
+
+        msg += "\n\n";
+        msg += &deps_to_bundle
+            .iter()
+            .map(|(package_id, (_, pseudo_extern_crate_name))| {
+                format!(
+                    "- `{}` as `crate::{}`\n",
+                    package_id, pseudo_extern_crate_name,
+                )
+            })
+            .join("");
+
+        let crates_available_on_atcoder = iproduct!(deps_to_bundle.keys(), ATCODER_CRATES)
+            .filter(|(id, s)| s.parse::<PkgSpec>().unwrap().matches(&metadata[id]))
+            .map(|(id, _)| format!("- `{}`\n", id))
+            .join("");
+
+        if !crates_available_on_atcoder.is_empty() {
+            msg += &format!(
+                "\nnote: attempted to bundle with the following crate(s), which are available on \
+                 AtCoder. to exclude them from bundling, run with `--exclude-atcoder-crates`\n\n{}",
+                crates_available_on_atcoder,
+            );
+        }
+
+        msg
+    };
+
     let code = bundle(
         &metadata,
         &bin_package,
         &bin,
-        &toolchain,
+        &deps_to_bundle,
         resolve_cfgs,
         &remove,
         minify,
         rustfmt,
         shell,
-    )?;
+    )
+    .with_context(|| error_message("could not bundle the code"))?;
 
     if check {
-        workspace::cargo_check_using_current_lockfile_and_cache(&metadata, &bin_package, &code)?;
+        workspace::cargo_check_using_current_lockfile_and_cache(
+            &metadata,
+            &bin_package,
+            &exclude,
+            &code,
+        )
+        .with_context(|| error_message("the bundled code was not valid"))?;
     }
 
     if let Some(output) = output {
@@ -213,23 +341,13 @@ fn bundle(
     metadata: &cm::Metadata,
     bin_package: &cm::Package,
     bin: &cm::Target,
-    toolchain: &str,
+    deps_to_bundle: &BTreeMap<&cm::PackageId, (&cm::Target, String)>,
     resolve_cfgs: bool,
     remove: &[Remove],
     minify: Minify,
     rustfmt: bool,
     shell: &mut Shell,
 ) -> anyhow::Result<String> {
-    let unused_deps = match cargo_udeps::cargo_udeps(&bin_package, &bin.name, &toolchain, shell) {
-        Ok(unused_deps) => unused_deps,
-        Err(warning) => {
-            shell.warn(warning)?;
-            hashset!()
-        }
-    };
-
-    let deps_to_bundle = metadata.deps_to_bundle(&bin_package.id, &unused_deps)?;
-
     let out_dirs =
         workspace::execute_build_scripts(metadata, deps_to_bundle.keys().copied(), shell)?;
 
@@ -246,7 +364,7 @@ fn bundle(
     let mut code = rust::process_extern_crate_in_bin(&code, |extern_crate_name| {
         matches!(
             metadata.dep_lib_by_extern_crate_name(&bin_package.id, extern_crate_name),
-            Ok(lib_package) if lib_package.is_available_on_atcoder_or_codingame()
+            Ok(lib_package) if deps_to_bundle.contains_key(&lib_package.id)
         )
     })?;
 
@@ -286,7 +404,10 @@ fn bundle(
             let content = rust::modify_macros(&content, &pseudo_extern_crate_name)?;
             let mut content = rust::insert_pseudo_extern_preludes(&content, &{
                 metadata
-                    .libs_with_extern_crate_names(&lib_package.id)?
+                    .libs_with_extern_crate_names(
+                        &lib_package.id,
+                        &deps_to_bundle.keys().copied().collect(),
+                    )?
                     .into_iter()
                     .map(|(package_id, extern_crate_name)| {
                         let (_, pseudo_extern_crate_name) =